test/ucx: extract GPU utilities into common header file
Move duplicated GPU wrapper functions into a shared gpu_utils.h header
to eliminate code duplication between test files.

Changes:
- Create test/unit/plugins/ucx/gpu_utils.h with common GPU utilities:
  * Error checking functions (checkCudaError, checkHipError)
  * GPU API wrappers (gpuSetDevice, gpuMalloc, gpuFree, gpuMemset)
  * Memory copy helper (gpuMemcpyD2H - only D2H is used)
  * Device count wrapper (gpuGetDeviceCount)
- Update ucx_backend_test.cpp to include gpu_utils.h
- Update ucx_worker_test.cpp to include gpu_utils.h
- Remove ~180 lines of duplicated code

This improves maintainability by having a single source of truth for
GPU API abstractions shared across UCX unit tests.
